% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_scale}
\alias{add_scale}
\title{Helper function to scale physicochemical properties}
\usage{
add_scale(df)
}
\arguments{
\item{df}{The dataframe with relevant property information in long form}
}
\description{
Helper function to scale physicochemical properties
}
