% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physchem.R
\name{create_pchem}
\alias{create_pchem}
\title{Generate physchem parameters from a set of molecular objects}
\usage{
create_pchem(source_pc, source_analogues, target_pc)
}
\arguments{
\item{source_pc}{Source analogue pchem.}

\item{source_analogues}{Source analogue identifiers}
}
\value{
A dataframe of physchem parameters for the set of source analogues for a specific target tagged by their dtxsid identifier and role
}
\description{
Generate physchem parameters from a set of molecular objects
}
