% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_sd}
\alias{filter_sd}
\title{Helper function to compute the SD threshold}
\usage{
filter_sd(df, n = 3)
}
\arguments{
\item{df}{A melted df with pchem properties derived from the summary_analysis}

\item{n}{The sd threshold where 3SD is the default}
}
\value{
A df with source analogues falling outside of the SD threshold
}
\description{
Helper function to compute the SD threshold
}
