% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem.R
\name{generate_fingerprints}
\alias{generate_fingerprints}
\title{Generate Morgan fingerprints for a set of substances characterised by their rcdk molecular objects}
\usage{
generate_fingerprints(substances)
}
\arguments{
\item{substances}{A dataframe of dtxsid and smiles}
}
\value{
A rcdk fingerprint object
}
\description{
Generate Morgan fingerprints for a set of substances characterised by their rcdk molecular objects
}
