% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem.R
\name{generate_fp}
\alias{generate_fp}
\title{Generate Morgan fingerprints for a set of substances characterised by their rcdk molecular objects}
\usage{
generate_fp(smiles, dtxsid)
}
\arguments{
\item{smiles}{smiles}
}
\value{
A rcdk fingerprint object
}
\description{
Generate Morgan fingerprints for a set of substances characterised by their rcdk molecular objects
}
