% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physchem.R
\name{generate_mol_obj}
\alias{generate_mol_obj}
\title{Generate mol objects from a dataframe of substances and their smiles}
\usage{
generate_mol_obj(smiles, dtxsid)
}
\arguments{
\item{substances}{A dataframe of dtxsid identifiers and smiles.}
}
\value{
A rcdk molecular object
}
\description{
Generate mol objects from a dataframe of substances and their smiles
}
