% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physchem.R
\name{generate_physchem}
\alias{generate_physchem}
\title{Generate physchem parameters from a set of molecular objects}
\usage{
generate_physchem(substances_mol)
}
\arguments{
\item{substances_mol}{A set of molecular objects.}
}
\value{
A dataframe of physchem parameters for the set of substances tagged by their dtxsid identifier
}
\description{
Generate physchem parameters from a set of molecular objects
}
