% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem.R
\name{get_analogues}
\alias{get_analogues}
\title{Identify candidate source analogues}
\usage{
get_analogues(sfp, tfp, k)
}
\arguments{
\item{sfp}{Source fingerprint object}

\item{tfp}{Target fingerprint object}

\item{k}{Number of source analogues to return}
}
\value{
A dataframe containing the source analogues and their pairwise Jaccard similarities
}
\description{
Identify candidate source analogues
}
