% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_df}
\alias{prep_df}
\title{Helper function to calculate similarity weighted activity}
\usage{
prep_df(df)
}
\arguments{
\item{df}{The df with physchem properties for target and analogues}
}
\description{
Helper function to calculate similarity weighted activity
}
