% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem.R
\name{process_substances}
\alias{process_substances}
\title{Preprocess csv file of substances and their identifiers}
\usage{
process_substances(substance_path)
}
\arguments{
\item{substance_path}{A file path and csv file name for a set of substances.}
}
\value{
A dataframe of dtxsid identifiers and smiles
}
\description{
Preprocess csv file of substances and their identifiers
}
