% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_array}
\alias{source_array}
\title{Get Neighbourhood List}
\usage{
source_array(all_analogues)
}
\arguments{
\item{all_analogues}{List of dataframes for each target and their associated pairwise similarities and dtxsid identifiers}
}
\value{
A vector of source analogues
}
\description{
Get Neighbourhood List
}
