% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_lst}
\alias{source_lst}
\title{Get Neighbourhood}
\usage{
source_lst(targets, source_fps, target_fps, k)
}
\arguments{
\item{targets}{A file of targets with dtxsid identifiers and smiles}

\item{source_fps}{Source fingerprint set}

\item{target_fps}{Target fingerprint set}
}
\value{
A list of analogues
}
\description{
Get Neighbourhood
}
