% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wtavg}
\alias{wtavg}
\title{Helper function to calculate similarity weighted activity}
\usage{
wtavg(id, df, outcome_col)
}
\arguments{
\item{id}{dtxsid}

\item{df}{The df with activity and similarities}

\item{outcome}{endpoint of interest}
}
\value{
A tibble with predicted value
}
\description{
Helper function to calculate similarity weighted activity
}
