% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_df}
\alias{prep_df}
\title{Helper function to calculate similarity weighted activity}
\usage{
prep_df(df)
}
\arguments{
\item{df}{A dataframe in long format containing the columns: \code{dtxsid}, \code{property}, \code{role}, and \code{property_value}}
}
\description{
Helper function to calculate similarity weighted activity
}
